ref(dotagents): Split agent runtime concerns
Move target definitions, MCP and hook writers, and subagent runtime handling into separate modules while keeping the agents barrel as a compatibility layer.

Update local agent instructions to describe the new source layout.
