+++ packages/dotagents/src/config/loader.ts
@@ -1,7 +1,7 @@
 import { readFile } from "node:fs/promises";
 import { parse as parseTOML } from "smol-toml";
 import { agentsConfigSchema, isWildcardDep, type AgentsConfig } from "./schema.js";
-import { allAgentIds } from "../agents/registry.js";
+import { allAgentIds } from "../targets/registry.js";
 import { applyDefaultRepositorySource, parseSource } from "@sentry/dotagents-lib";
 
 export class ConfigError extends Error {
