+++ packages/dotagents/src/targets/hook-writer.ts
