+++ packages/dotagents/src/targets/mcp-writer.ts
