+++ packages/dotagents/src/targets/paths.test.ts
