+++ packages/dotagents/src/targets/paths.ts
