+++ packages/dotagents/src/targets/registry.test.ts
