+++ packages/dotagents/src/targets/registry.ts
